Move FFE fixtures under allowed utils package
